requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
pydantic==2.5.0
pyresparser
python-dotenv==1.0.0
azure-ai-inference==1.0.0b3
azure-core==1.29.5
pdfplumber==0.9.0
python-docx==0.8.11
python-multipart==0.0.6
email-validator==2.1.0